create fakedb_test.go and ranme db_spy_test.go to spydb_test.go
