+++ spydb_test.go
